- Updating all scripts to use new approach

+++ confluent-sigma/sigma-streams/test/bin/produce-data.sh
@@ -6,12 +6,33 @@
 # Parameter 2 the topic name
 # Parameter 3 the number of partitions to for the topic
 
-source bin/auto-configure.sh
+
+# These scripts are intended to be run from the test directory
+
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+
+echo "Script dir is $SCRIPT_DIR"
+
+if [ -f "$SCRIPT_DIR/../../bin/auto-configure.sh" ] ; then
+  source $SCRIPT_DIR/../../bin/auto-configure.sh
+fi
+
+
+if [ ! -f $SIGMA_PROPS ] ; then
+  echo "sigma properties not found.  Suggested path from auto-configure is $SIGMA_PROPS"
+  exit
+fi
+
+
+bootstrap_key="bootstrap.server"
+BOOTSTRAP_SERVER=$(grep "^$bootstrap_key=" "$SIGMA_PROPS" | cut -d'=' -f2-)
+
+
 
 docker run -v $(pwd)/test/config:/mnt/config --rm --network=host confluentinc/cp-server:latest \
   kafka-topics \
-  --bootstrap-server ${CCLOUD_BOOTSTRAP_SERVER} \
-  --command-config /mnt/config/java.config \
+  --bootstrap-server ${BOOTSTRAP_SERVER} \
+  --command-config $SIGMA_PROPS \
   --topic test \
   --create \
   --replication-factor 3 \
